Fix some date parsing/sorting issues and separate parsing and requesting into to separate classes to adhere to single responsibility principle
